# api/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Go CVEs API
  version: "0.1"
servers:
- url: /api/v1/
paths:
  /insert:
    post:
      tags:
      - Insert Record
      summary: Insert record into the database
      requestBody:
        description: Record data
        content:
          application/json:
            schema:
              type: object
              oneOf:
                - $ref: '#/components/schemas/Commit'
                - $ref: '#/components/schemas/Issue'
                - $ref: '#/components/schemas/Pullrequest'
                - $ref: '#/components/schemas/ProbableCVE'
              example:
                type: repository | issue | commit | pullrequest | probablecve
                object: object of one of the above type
        required: true
      responses:
        200:
          description: Data inserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResponse'
        400:
          description: Bad request from the client
          content: {}
        500:
          description: Internal server error
          content: {}
  /update:
    post:
      tags:
      - Update Record
      summary: update github issues, commits, PR and probable cves
      requestBody:
        description: data for the object to update
        content:
          application/json:
            schema:
              type: object
              oneOf:
                - $ref: '#/components/schemas/Commit'
                - $ref: '#/components/schemas/Issue'
                - $ref: '#/components/schemas/Pullrequest'
                - $ref: '#/components/schemas/ProbableCVE'
              example:
                type: repository | issue | commit | pullrequest | probablecve
                id: id of the object to update
                object: object of one of the above type
        required: true
      responses:
        200:
          description: CVEs updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResponse'
        400:
          description: Bad request from the client
          content: {}
        500:
          description: Internal server error
          content: {}
  /delete:
    post:
      tags:
      - Delete Record
      summary: delete github issues, commits, PR and probable cves
      requestBody:
        description: data for the object to delete
        content:
          application/json:
            schema:
              type: object
              oneOf:
                - $ref: '#/components/schemas/Commit'
                - $ref: '#/components/schemas/Issue'
                - $ref: '#/components/schemas/Pullrequest'
                - $ref: '#/components/schemas/ProbableCVE'
              example:
                type: repository | issue | commit | pullrequest | probablecve
                object: object of one of the above type
        required: true
      responses:
        200:
          description: object deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResponse'
        400:
          description: Bad request from the client
          content: {}
        500:
          description: Internal server error
          content: {}
  /getid:
    post:
      tags:
      - Get ID
      summary: Get ID of the object
      requestBody:
        description: data for the object to get ID for
        content:
          application/json:
            schema:
              type: object
              oneOf:
                - $ref: '#/components/schemas/Commit'
                - $ref: '#/components/schemas/Issue'
                - $ref: '#/components/schemas/Pullrequest'
                - $ref: '#/components/schemas/ProbableCVE'
              example:
                type: repository | issue | commit | pullrequest | probablecve
                object: object of one of the above type
        required: true
      responses:
        200:
          description: object found
          content:
            application/json:
              schema:
                type: object
                example:
                  id: ID of the object
        400:
          description: Bad request from the client
          content: {}
        500:
          description: Internal server error
          content: {}
  /get:
    get:
      tags:
      - Get
      summary: Get an object by it's ID
      parameters:
      - name: type
        in: query
        description: type of the object to query
        required: true
        schema:
          type: string
          enum: ['repository', 'issue', 'pullrequest', 'probablecve', 'commit']
      - name: id
        in: query
        description: ID of the object to fetch
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Object found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Commit'
                  - $ref: '#/components/schemas/Issue'
                  - $ref: '#/components/schemas/Pullrequest'
                  - $ref: '#/components/schemas/ProbableCVE'
                example:
                  id: id of the object
                  details: details of the object
        400:
          description: Bad request from the client
          content: {}
        404:
          description: Data not found
          content: {}
        500:
          description: Internal server error
          content: {}
  /getcves:
    get:
      tags:
      - Get CVE
      summary: Get CVEs for a given date interval
      parameters:
      - name: start-date
        in: query
        description: start date of the interval in dd-mm-yyyy format
        required: true
        schema:
          type: string
      - name: end-date
        in: query
        description: end date of the interval in dd-mm-yyyy format
        required: true
        schema:
          type: string
      responses:
        200:
          description: Probable CVEs for the given date interval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbableCVEArray'
        400:
          description: Bad request from the client
          content: {}
        404:
          description: Data not found
          content: {}
        500:
          description: Internal server error
          content: {}
components:
  schemas:
    ProbableCVE:
      title: Probable CVE
      type: object
      properties:
        id:
          type: string
        cve_id:
          type: string
        ecosystem:
          type: string
        cause_type:
          type: string
        cause_id:
          type: integer
          description: id of issue, pr, commit that caused this CVE
        cve_date:
          type: string
        confidence_score:
          type: number
        identified_date:
          type: string
        review_comments:
          type: string
        review_status:
          type: integer
          description: 0 means not reviewed, 1 means reviewed
        is_cve:
          type: integer
          description: 0 for not a cve, 1 means a cve
        is_triaged_internally:
          type: integer
          description: 0 for not triaged internally, 1 means triaged
        last_updated_at:
          type: string
        reviewed_by:
          type: string
      description: Probable CVEs
    UpdateResponse:
      title: Update response
      type: object
      properties:
        message:
          type: string
      description: response for the update cve call
    Issue:
      title: Issue
      type: object
      properties:
        number:
          type: integer
          description: GH issue number in the upstream
        created_at:
          type: string
          description: creation date of the issue
        updated_at:
          type: string
          description: last updated date of the issue
        body:
          type: string
          description: body of the issue
        title:
          type: string
          description: title of the issue
        repo_id:
          type: integer
          description: repository id
        url:
          type: string
          description: url of the issue
        closed_at:
          type: string
          description: closing date of the issue
        is_security:
          type: integer
          description: binary flag indicating if the issue is a security issue
        is_probable_cve:
          type: integer
          description: binary flag indicating if the issue is a probable cve
        is_passed_through_model:
          type: integer
          description: binary flag indicating if the issue is passed through model
            or not
      description: Github Issue
    Commit:
      title: Commit
      type: object
      properties:
        sha:
          type: string
          description: SHA of the commit
        created_at:
          type: string
          description: timestamp at which the commit was created
        message:
          type: string
          description: commit message
        repo_id:
          type: string
          description: Repository id from the repositories table
        url:
          type: string
          description: URL of the commit
        is_security:
          type: string
          description: binary flag showing if commit is security or not
        is_probable_cve:
          type: string
          description: binary flag showing if commit is probable cve or not
        is_passed_through_model:
          type: string
          description: binary flag showing if commit is passed through model or not
      description: Github commits
    ProbableCVEArray:
      title: Probable CVEs in Array
      type: object
      properties:
        probablecves:
          type: array
          items:
            $ref: '#/components/schemas/ProbableCVE'
      description: Probable CVEs in Array
    Pullrequest:
      title: Pull request
      type: object
      properties:
        number:
          type: integer
          description: Pull request number
        created_at:
          type: string
          description: creation date of the pull request
        updated_at:
          type: string
          description: updation date of the pull request
        merged_at:
          type: string
          description: merged date of the pull request
        closed_at:
          type: string
          description: closing date of the pull request
        title:
          type: string
          description: title of the PR
        body:
          type: string
          description: body of the PR
        repo_id:
          type: string
          description: Repository id from the repositories table
        url:
          type: string
          description: URL for the PR
        patch_url:
          type: string
          description: Patch URL of the PR
        is_security:
          type: integer
          description: binary flag showing PR is security related or not
        is_probable_cve:
          type: integer
          description: binary flag showing PR is probable cve or not
        is_passed_through_model:
          type: integer
          description: binary flag showing if PR is scored through model or not
      description: Github Pull Request